# openapi.yaml
openapi: 3.1.0
info:
  title: komondor-api
  version: "1.0.0"
  summary: TSL Sequence Data API — projects, samples, sequencing runs and their files.
  description: |
    A description of what komondor-api **does today**, derived by reading
    `routes/` and `models/`. It is a record of current behaviour, not a
    proposal. Where the current behaviour is a known defect it is documented as
    it is, and the operation carries a `DRIFT` or `DEFECT` note saying so.
    Fixing one of those is a change to the API and must update this file in the
    same commit.

    `__tests__/contract/openapi.test.js` fails if a route exists in `routes/`
    with no entry here, so the two cannot silently diverge.

    ## Known drift from the sibling clients

    Three repositories consume this API: `komondor-web` (Nuxt 2 UI),
    `komondor-power` (bulk CSV ingest) and `komondor-nudge` (a scheduled
    worker that reads the same MongoDB directly rather than over HTTP). The
    following are the places where their assumptions and this server's
    behaviour have come apart. See `docs/CONTRACTS.md` for the policy.

    1. **Routes are mounted at the ROOT, not under `/api`.** `app.js` calls
       `app.use(authRoutes)` with no path prefix, so login is `POST /login`,
       not `POST /api/login` or `POST /auth/login`. The comment above the login
       handler in `routes/auth.js` still reads "Add POST - /api/login", which
       is where the belief came from. `komondor-power`'s client tries
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /directory-files:
    get:
      tags: [Files]
      summary: List a directory inside the HPC transfer area.
      description: |
        DRIFT — FAILURE IS REPORTED AS **200** with an `error` body. Missing
        parameter, unconfigured `HPC_TRANSFER_DIRECTORY`, non-existent
        directory, and *empty directory* all answer 200 with
        `{ "error": "..." }`. This is deliberately preserved
        (BREAKING_CHANGES §9) because consumers detect failure by the presence
        of `error`, not by status code.

        The one exception is a rejected path, which answers a real **403**.
        Clients therefore need both branches — see `utils/apiError.js` in
        komondor-web.

        Note that an empty directory is an error here, not an empty list.

        Path containment uses `resolveBelow`, so a name normalising back to the
        root (`.`, `./`, `a/..`) is refused rather than listing every group's
        inbound directory.
      parameters:
        - name: targetDirectoryName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            Either the listing OR a failure. Exactly one of `filesResults` and
            `error` is present.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [filesResults]
                    properties:
                      filesResults:
                        type: array
                        items:
                          type: string
                  - type: object
                    required: [error]
                    properties:
                      error:
                        type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          description: The path escaped the transfer directory, or resolved to its root.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    const: "Access denied: Invalid directory path"
  /directory-files/debug:
    get:
      tags: [Files]
      summary: Report how a directory name resolves on disk. Admin only.
      description: |
        Admin-gated because the response exposes server filesystem layout and
        configuration — `process.cwd()`, `__dirname` and the value of
        `HPC_TRANSFER_DIRECTORY` (BREAKING_CHANGES §8).

        `targetDirectoryName` defaults to the string `"cheese"` when omitted.
      parameters:
        - name: targetDirectoryName
          in: query
          required: false
          schema:
            type: string
            default: cheese
      responses:
        "200":
          description: Resolution report.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cwd:
                    type: string
                  __dirname:
                    type: string
                  HPC_TRANSFER_DIRECTORY:
                    type: [string, "null"]
                  targetDirectoryName:
                    type: [string, "null"]
                  cleanedTargetDirectoryName:
                    type: [string, "null"]
                  dirRoot:
                    type: [string, "null"]
                    description: "`null` when the path would escape the transfer directory."
                  withinTransferDirectory:
                    type: boolean
                  exists:
                    type: boolean
                  isDirectory:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AdminRequired"
  /directory-files/verify-md5:
    post:
      tags: [Files]
      summary: Checksum a file in the transfer directory against an expected MD5.
      description: |
        UNUSUAL RESPONSE FRAMING. Headers are flushed with status 200 before
        the checksum starts, and a space character is written periodically to
        stop a reverse proxy timing out on a large file. The JSON body is
        appended at the end.

        Consequences for clients: the status code is 200 before the outcome is
        known, so a mid-stream failure arrives as 200 with an `error` body; and
        the body has leading whitespace, so it must be trimmed before parsing.

        `expectedMd5` is lower-cased and trimmed before comparison.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [directoryName, fileName, expectedMd5]
              properties:
                directoryName:
                  type: string
                fileName:
                  type: string
                expectedMd5:
                  type: string
      responses:
        "200":
          description: |
            The comparison, or a mid-stream failure. Preceded by padding
            whitespace in both cases.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [fileName, expectedMd5, calculatedMd5, matches]
                    properties:
                      fileName:
                        type: string
                      expectedMd5:
                        type: string
                      calculatedMd5:
                        type: string
                      matches:
                        type: boolean
                  - type: object
                    required: [error]
                    properties:
                      error:
                        type: string
                      requestId:
                        type: string
        "400":
          description: A required field is missing or not a string.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  requestId:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          description: Path traversal attempt.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    const: "Access denied: Invalid file path"
                  requestId:
                    type: string
        "404":
          description: No such file, or the path is not a regular file.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  requestId:
                    type: string
        "500":
          description: Failed before headers were sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  requestId:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <token>`, signed with `JWT_SECRET` (HS256) by
        `lib/utils/jwtSign.js`.

        **The `exp` claim is mandatory.** `lib/utils/getUserFromRequest.js`
        rejects a structurally valid, correctly signed token that carries no
        `exp` by raising `TokenExpiredError`, which surfaces as 401. This was
        added deliberately (commit "refuse legacy tokens without an expiry
        claim"): tokens issued before August 2026 had no expiry and stayed
        valid forever, including tokens minted during the LDAP incident with an
        empty `groups` claim. Such tokens can only be fixed by re-logging in.

        Tokens expire after `JWT_EXPIRES_IN`, default `7d`.

        **Group membership is a snapshot.** `groups` is resolved once, at
        login, and never re-read. A user added to a group sees nothing new
        until their token expires and they sign in again — that expiry is the
        upper bound on the staleness.

        Requests with NO `Authorization` header are not rejected by the
        app-level middleware; they simply arrive with no `req.user`, and each
        route's `isAuthenticated` guard turns that into a 401. Routes without
        that guard (`/projects/names`, `GET /options/*`, `/health`, `/test`,
        `/login`, `/logout`, `/uploads`) serve anonymous callers.
...
  responses:
    Unauthenticated:
      description: |
        No usable token. Two distinct bodies reach clients as 401:
        `{ error: "Authentication required" }` from a route's
        `isAuthenticated` guard, and
        `{ error: "Invalid or expired authentication token", detail }` from the
        app-level middleware in `app.js` when a token was present but bad or
        expired.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  error:
                    const: Authentication required
              - type: object
                properties:
                  error:
                    const: Invalid or expired authentication token
                  detail:
                    type: string
                    description: The underlying jsonwebtoken message.
                  requestId:
                    type: string
                    description: Present only when raised by the terminal error handler.
    AdminRequired:
      description: The caller's token lacks `isAdmin`.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                const: Admin access required
...